+++ core-face-detect/client.js
@@ -55,7 +55,7 @@ function displayError(text) {
 async function detectFace(blob) {
   const requestData = await buildRequestData(blob);
 
-  const response = await fetch("/api/v6/face/detect", {
+  const response = await fetch("/core/api/v6/face/detect", {
     method: "POST",
     body: JSON.stringify(requestData),
     headers: {
